# docs/swagger.yaml
basePath: /
definitions:
  handler.PaginatedResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Task'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  handler.createTaskRequest:
    properties:
      done:
        example: false
        type: boolean
      title:
        example: Buy groceries
        type: string
    type: object
  handler.updateTaskRequest:
    properties:
      done:
        example: true
        type: boolean
      title:
        example: Buy milk
        type: string
    type: object
  model.Task:
    properties:
      done:
        type: boolean
      id:
        type: integer
      title:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://example.com/support
  description: Simple REST API for managing tasks with pagination and in-memory cache
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://example.com/terms/
  title: GoSimpleRestApi
  version: 1.0.0
paths:
  /tasks:
    get:
      description: Retrieve paginated list of tasks from the database
      parameters:
      - default: 0
        description: 'Offset for pagination (default: 0)'
        in: query
        name: offset
        type: integer
      - default: 20
        description: 'Limit per page (default: 20, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.PaginatedResponse'
        "400":
          description: Invalid pagination parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Database error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all tasks with pagination
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Add a new task to the database
      parameters:
      - description: Task data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.createTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Task'
        "400":
          description: Missing title or invalid JSON
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Database error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new task
      tags:
      - tasks
  /tasks/{id}:
    delete:
      description: Remove a task by ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Task not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Database error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a task
      tags:
      - tasks
    put:
      consumes:
      - application/json
      description: Modify a task by ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated task data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.updateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Task'
        "400":
          description: Invalid ID or request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Task not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Database error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update an existing task
      tags:
      - tasks
swagger: "2.0"